requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
pydantic
pytest
python-dotenv
